[Refactor] Aside 구조 재정리 + hooks 폴더 생성(#61)
