+++ src/components/aside/AsideMenu/Aside.jsx
@@ -1,19 +1,19 @@
 import { useState } from 'react';
 import './Aside.css';
 import { useCookies } from 'react-cookie';
-import tagIcon from '../../assets/icons/tag-icon.png';
-import chatIcon from '../../assets/icons/chat-icon.png';
-import personIcon from '../../assets/icons/person-icon.png';
-import fileIcon from '../../assets/icons/file-icon.png';
-import writingIcon from '../../assets/icons/writing-icon.png';
-import mobileAsideIcon from '../../assets/icons/mobile-aside-icon.png';
+import tagIcon from '../../../assets/icons/tag-icon.png';
+import chatIcon from '../../../assets/icons/chat-icon.png';
+import personIcon from '../../../assets/icons/person-icon.png';
+import fileIcon from '../../../assets/icons/file-icon.png';
+import writingIcon from '../../../assets/icons/writing-icon.png';
+import mobileAsideIcon from '../../../assets/icons/mobile-aside-icon.png';
 import { Link, useNavigate } from 'react-router-dom';
-import useCurrentPath from '../../hooks/useCurrentPath';
-import useLastCommentStore from '../../store/useLastCommenStore';
-import useShowExtend from '../../hooks/useShowExtend';
-import WritingMenu from './WritingMenu/WritingMenu';
-import useIsMobile from '../../hooks/useIsMobile';
-import FileMenu from './FileMenu/FileMenu';
+import useCurrentPath from '../../../hooks/useCurrentPath';
+import useLastCommentStore from '../../../store/useLastCommenStore';
+import useShowExtend from '../AsideMenu/hooks/useShowExtend';
+import WritingMenu from '../WritingMenu/WritingMenu';
+import useIsMobile from '../../../hooks/useIsMobile';
+import FileMenu from '../FileMenu/FileMenu';
 
 function Aside() {
   const [cookies] = useCookies(['studentId']);
